Wrap `stream.Read` for safety consideration
`FrameReader` and `FrameWriter` has code that reads from the Stream.
This code uses `Stream.Read` which does not guarantee that requested
number of bytes are read and it does not check how much bytes it read.

Simple solution would to be replace it with `Stream.ReadExactly`, but it
is not supported in net6.

So, only choice here is implement a code that blocks until requested
bytes are read.